+++ reikna/transformations.py
@@ -9,11 +9,34 @@ from reikna.core import Transformation, Parameter, Annotation, Type
 
 def copy(arr_t, out_arr_t=None):
     """
+    Returns an identity transformation (1 output, 1 input): ``output = input``.
+    Output array type ``out_arr_t`` may have different strides,
+    but must have the same shape and data type.
+    """
+    if out_arr_t is None:
+        out_arr_t = arr_t
+    else:
+        if out_arr_t.shape != arr_t.shape or out_arr_t.dtype != arr_t.dtype:
+            raise ValueError("Input and output arrays must have the same shape and data type")
+
+    return Transformation(
+        [Parameter('output', Annotation(out_arr_t, 'o')),
+        Parameter('input', Annotation(arr_t, 'i'))],
+        "${output.store_same}(${input.load_same});")
+
+
+def copy_broadcasted(arr_t, out_arr_t=None):
+    """
     Returns an identity transformation (1 output, 1 input): ``output = input``,
     where ``input`` may be broadcasted (with the same semantics as ``numpy.broadcast_to()``).
     Output array type ``out_arr_t`` may have different strides,
     but must have compatible shapes the same shape and data type.
+
+    .. note::
+
+        This is an input-only transformation.
     """
+
     if out_arr_t is None:
         out_arr_t = arr_t
 
@@ -41,7 +64,8 @@ def copy(arr_t, out_arr_t=None):
             %endif
         %endfor
         ));
-        """)
+        """,
+        connectors=['output'])
 
 
 def cast(arr_t, dtype):